A bit more tweaks
