+++ builtin/providers/azurerm/resource_arm_express_route_circuit.go
@@ -1,12 +1,15 @@
 package azurerm
 
 import (
+	"bytes"
 	"log"
+	"strings"
 
 	"fmt"
 
 	"github.com/Azure/azure-sdk-for-go/arm/network"
 	"github.com/hashicorp/errwrap"
+	"github.com/hashicorp/terraform/helper/hashcode"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/helper/validation"
 )
@@ -201,10 +204,36 @@ func resourceArmExpressRouteCircuitDelete(d *schema.ResourceData, meta interface
 	}
 
 	_, err = ercClient.Delete(resGroup, name, make(chan struct{}))
-	if err != nil {
-		return errwrap.Wrapf("Error Deleting ExpressRouteCircuit {{err}}", err)
-	}
+	return err
+}
 
-	d.SetId("")
-	return nil
+func expandExpressRouteCircuitSku(skuSettings *schema.Set) *network.ExpressRouteCircuitSku {
+	v := skuSettings.List()[0].(map[string]interface{}) // [0] is guarded by MinItems in schema.
+	tier := v["tier"].(string)
+	family := v["family"].(string)
+	name := fmt.Sprintf("%s_%s", tier, family)
+
+	return &network.ExpressRouteCircuitSku{
+		Name:   &name,
+		Tier:   network.ExpressRouteCircuitSkuTier(tier),
+		Family: network.ExpressRouteCircuitSkuFamily(family),
+	}
+}
+
+func flattenExpressRouteCircuitSku(sku *network.ExpressRouteCircuitSku) []interface{} {
+	return []interface{}{
+		map[string]interface{}{
+			"tier":   string(sku.Tier),
+			"family": string(sku.Family),
+		},
+	}
+}
+
+func resourceArmExpressRouteCircuitSkuHash(v interface{}) int {
+	var buf bytes.Buffer
+	m := v.(map[string]interface{})
+	buf.WriteString(fmt.Sprintf("%s-", strings.ToLower(m["tier"].(string))))
+	buf.WriteString(fmt.Sprintf("%s-", strings.ToLower(m["family"].(string))))
+
+	return hashcode.String(buf.String())
 }
